+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## master
+
+* [bugfix] Capture password without echo only if terminal is present
+
 ## 0.4.28 / 2014-03-25
 
 * [bugfix] Fix password no_echo input on Ruby 1.9